Add files via upload
This is the game called Where's Waldo. Here we will be having a certain character called Waldo in the picture wallpaperflare.com_wallpaper(3).jpg. So whenever we click on this character we get 100 points. To improvise the game, I took four people as targets (Two waldo's and two red and white sweaters). The game is of 400 points as of now. We can take new objects, set them as targets, and increase the set of numbers.